+++ src/realm/util/serializer.hpp
@@ -37,6 +37,7 @@ class StringData;
 class Timestamp;
 class LinkMap;
 class UUID;
+class TypeOfValue;
 enum class ExpressionComparisonType : unsigned char;
 
 namespace util {
@@ -68,6 +69,8 @@ template <>
 std::string print_value<>(realm::ObjKey);
 template <>
 std::string print_value<>(realm::UUID);
+template <>
+std::string print_value<>(realm::TypeOfValue);
 
 // General implementation for most types
 template <typename T>
@@ -87,6 +90,8 @@ std::string print_value(Optional<T> value)
         return "NULL";
     }
 }
+
+StringData get_printable_table_name(StringData name);
 
 struct SerialisationState {
     std::string describe_column(ConstTableRef table, ColKey col_key);